Numerical survival function

Constructs the numerical survival function from a probability density function.

survival_function(pdf)

Arguments

pdf

The probability density function.

Value

Returns the numerical value of survival at time \(t\).

Details

The function survival_function() returns a function in terms of \(t\) and the additional parameters of the probability density function that is passed as an argument to pdf. The returned function calculates survival numerically, making the implementation of survival functions quick. The returned function is vectorized in \(t\), that is, a vector of time instances can be passed as an argument.

The survival function returned by survival_function() also has a t0 argument that defaults to t0 = 0. The t0 argument is the initial time for the calculation of the survival function. The t0 argument is useful for calculating the survival function from a start time different from zero. In most cases, you should not change the default value of t0.

See also

Examples

survival_weibull <- survival_function(dweibull)
survival_weibull(0:10, shape = 2, scale = 1)
#>  [1] 1.000000e+00 3.678794e-01 1.831564e-02 1.234098e-04 1.125352e-07
#>  [6] 1.392675e-11 0.000000e+00 1.110223e-16 0.000000e+00 5.218048e-15
#> [11] 3.885781e-14
#> attr(,"time")
#>  [1]  0  1  2  3  4  5  6  7  8  9 10
#> attr(,"class")
#> [1] "survival_function"